==> 02_UKB/01_Load_data/00_load_demographics_and_merge_label.ipynb <==
import pandas as pd
import os
import subprocess

def download_data(outputfolder = "/opt/notebooks/00_data_for_loop/"):
    mri_tables = os.path.join(outputfolder, "mri_tables/")
    os.makedirs(mri_tables, exist_ok=True)
    command = f"dx download -r project-GvJ4qQ8Jk6k7QFZ0Jzjp7vJ6:/Users/Esra_Lenz/02_Data/00_mri_tables/03_combined_mris/00_confounded/ --output '{mri_tables}'"
    subprocess.run(command, shell=True,capture_output=True, text=True)

    labels = os.path.join(outputfolder, "labels")
    os.makedirs(labels, exist_ok=True)
    command = f"dx download -r project-GvJ4qQ8Jk6k7QFZ0Jzjp7vJ6:/Users/Esra_Lenz/02_Data/01_labels/merged_multitarget_df_with_demographics.csv --output '{labels}'"
    subprocess.run(command, shell=True,capture_output=True, text=True)

download_data()

mri_data = "/opt/notebooks/00_data_for_loop/mri_tables/"
#load specific files to ../00_data/mris if they contain 
output_dir = "../00_data/mris/"
os.makedirs(output_dir, exist_ok=True)
files_to_find = ["aparc.volume.csv", "aparc.thickness.csv", "aseg.volume.csv"]
for root, dirs, files in os.walk(mri_data):
    for file in files:
        for file_to_find in files_to_find:
            if file == file_to_find:
                print(file)
                df = pd.read_csv(os.path.join(root, file))
                #chang ending to csv
                name = file.replace(".tsv", ".csv")
                df.to_csv(os.path.join(output_dir, file), index=False)

merged_df = pd.read_csv("/opt/notebooks/00_data_for_loop/labels/merged_multitarget_df_with_demographics.csv")

merged_df


demographics_covariates = merged_df[["ID", "sex", "age_at_assessment", "assessment_centre"]]

#rename demographics_covariates to demographics_covariates
demographics_covariates.rename(
    columns={"ID": "ID", "sex": "Sex", "assessment_centre": "Site", "age_at_assessment": "Age"},
    inplace=True,
)

os.makedirs("../00_data/age_label/", exist_ok=True)
demographics_covariates.to_csv("../00_data/age_label/all_ages_all_ids.csv", index=False)

demographics_covariates.to_csv("../00_data/UKB_demographics_covariates.csv", index=False)

demographics_covariates

#download validation data
os.makedirs("../00_data/validation_data/", exist_ok=True)
command = "dx download -r project-GqzxkVQJ34g6ygFJ4ZbvqBYF:/Esra/00_CLIP/03_Validation_Data/00_National_Cohort/ --output '../00_data/validation_data/'"
subprocess.run(command, shell=True,capture_output=True, text=True)
