==> 02_UKB/01_Load_data/01_subset_age.ipynb <==
import pandas as pd

df = pd.read_csv("../00_data/UKB_demographics_covariates.csv")

print(df["Age"].describe())


df["Age"].hist()

df["Sex"].value_counts()

#box blot
df.boxplot(column="Age")

#just people from 18 to 26
df_middle = df[df["Age"] <= 74]
len(df_middle)
# didnt do that yet Could be an experiment
